chore: put validators in their own files
It will allow us to use them in other models
