+++ backend/api/db/models/csv.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class CsvFieldError(BaseModel):
+    key: str | None = None
+    error: str
